task-2

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,14 +1,14 @@
 #include "lists.h"
 
 /**
- * listint_t - returns the number of elements in a linked list
+ * listint_len - returns the number of elements in a linked list
  * @h:pointer to the header node
  *
  * Return:number of elements
  */
 size_t listint_len(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
